requirements.txt
-i https://pypi.org/simple
bottle==0.12.17
gevent==1.5a1
greenlet==0.4.15 ; platform_python_implementation == 'CPython'
pyserial==3.4
